chore: rename lib/common/stdostream.h => lib/common/StdOStream.h
